+++ assets/sass/app.scss
@@ -1,11 +1,13 @@
-// Add tailwindcss base
+/*! purgecss start ignore */
 @tailwind base;
+/*! purgecss end ignore */
 
 // Add our custom base styles
 @import "base/base";
 
-// Add tailwindcss components
+/*! purgecss start ignore */
 @tailwind components;
+/*! purgecss end ignore */
 
 // Add custom components
 @import "components/components";
