+++ assets/sass/vendors/_vendors.scss
@@ -1,0 +1,3 @@
+/*! purgecss start ignore */
+@import "wordpress";
+/*! purgecss end ignore */
