+++ package.json
@@ -14,7 +14,6 @@
   "devDependencies": {
     "@babel/core": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
-    "@fullhuman/postcss-purgecss": "^1.3.0",
     "autoprefixer": "^9.7.4",
     "babel-loader": "^8.0.6",
     "concurrently": "^5.1.0",
@@ -25,7 +24,7 @@
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
     "sass-loader": "^7.1.0",
-    "tailwindcss": "^1.2.0",
+    "tailwindcss": "^1.4.1",
     "terser-webpack-plugin": "^2.3.4",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10"
