+++ tailwind.config.js
@@ -1,5 +1,50 @@
+const whitelist = []
+
+const whitelistPatterns = [
+  /^home(-.*)?$/,
+  /^blog(-.*)?$/,
+  /^archive(-.*)?$/,
+  /^date(-.*)?$/,
+  /^error404(-.*)?$/,
+  /^admin-bar(-.*)?$/,
+  /^search(-.*)?$/,
+  /^nav(-.*)?$/,
+  /^wp(-.*)?$/,
+  /^screen(-.*)?$/,
+  /^navigation(-.*)?$/,
+  /^(.*)-template(-.*)?$/,
+  /^(.*)?-?single(-.*)?$/,
+  /^postid-(.*)?$/,
+  /^post-(.*)?$/,
+  /^attachmentid-(.*)?$/,
+  /^attachment(-.*)?$/,
+  /^page(-.*)?$/,
+  /^(post-type-)?archive(-.*)?$/,
+  /^author(-.*)?$/,
+  /^category(-.*)?$/,
+  /^tag(-.*)?$/,
+  /^menu(-.*)?$/,
+  /^tags(-.*)?$/,
+  /^tax-(.*)?$/,
+  /^term-(.*)?$/,
+  /^date-(.*)?$/,
+  /^(.*)?-?paged(-.*)?$/,
+  /^depth(-.*)?$/,
+  /^children(-.*)?$/,
+  /^h[1-6]?$/,
+]
+
 module.exports = {
+  purge: {
+    content: ['**/*.php'],
+    options: {
+      whitelist,
+      whitelistPatterns,
+    }
+  },
+  
   theme: {
+
     extend: {},
 
     fontFamily: {
@@ -17,3 +62,5 @@ module.exports = {
 
   plugins: [],
 }
+
+
